# Test performance improvements