--- a/detekt-api/src/test/kotlin/io/gitlab/arturbosch/detekt/api/EntitySpec.kt
+++ b/detekt-api/src/test/kotlin/io/gitlab/arturbosch/detekt/api/EntitySpec.kt
@@ -7,6 +7,7 @@ import org.jetbrains.kotlin.psi.KtNamedFunction
 import org.jetbrains.kotlin.psi.psiUtil.collectDescendantsOfType
 import org.jetbrains.kotlin.psi.psiUtil.findDescendantOfType
 import org.spekframework.spek2.Spek
+import org.spekframework.spek2.lifecycle.CachingMode
 import org.spekframework.spek2.style.specification.describe
 import java.nio.file.Paths
 
@@ -14,7 +15,8 @@ class EntitySpec : Spek({
 
     describe("entity signatures") {
 
-        val code = compileContentForTest("""
+        val code by memoized(CachingMode.SCOPE) {
+            compileContentForTest("""
             package test
 
             class C : Any() {
@@ -24,10 +26,11 @@ class EntitySpec : Spek({
 
             fun topLevelFun(number: Int) = Unit
         """.trimIndent(), Paths.get("/full/path/to/Test.kt").toString())
+        }
 
         describe("functions") {
 
-            val functions = code.collectDescendantsOfType<KtNamedFunction>()
+            val functions by memoized { code.collectDescendantsOfType<KtNamedFunction>() }
 
             it("includes full function header, class name and filename") {
                 val memberFunction = functions.first { it.name == "memberFun" }