--- a/detekt-core/build.gradle.kts
+++ b/detekt-core/build.gradle.kts
@@ -10,4 +10,5 @@ dependencies {
 
     testImplementation(project(":detekt-rules"))
     testImplementation(project(":detekt-test"))
+    testImplementation(testFixtures(project(":detekt-api")))
 }