--- a/detekt-core/src/test/kotlin/io/gitlab/arturbosch/detekt/core/rules/SingleRuleProviderSpec.kt
+++ b/detekt-core/src/test/kotlin/io/gitlab/arturbosch/detekt/core/rules/SingleRuleProviderSpec.kt
@@ -16,23 +16,25 @@ internal class SingleRuleProviderSpec : Spek({
 
     describe("SingleRuleProvider") {
 
-        val provider = SingleRuleProvider(
-            "MagicNumber",
-            object : RuleSetProvider {
-                override val ruleSetId: String = "style"
-                override fun instance(config: Config): RuleSet {
-                    val rule = object : Rule(config) {
-                        override val issue = Issue(
-                            "MagicNumber",
-                            Severity.CodeSmell,
-                            "",
-                            Debt.FIVE_MINS
-                        )
+        val provider by memoized {
+            SingleRuleProvider(
+                "MagicNumber",
+                object : RuleSetProvider {
+                    override val ruleSetId: String = "style"
+                    override fun instance(config: Config): RuleSet {
+                        val rule = object : Rule(config) {
+                            override val issue = Issue(
+                                "MagicNumber",
+                                Severity.CodeSmell,
+                                "",
+                                Debt.FIVE_MINS
+                            )
+                        }
+                        return RuleSet(ruleSetId, listOf(rule))
                     }
-                    return RuleSet(ruleSetId, listOf(rule))
                 }
-            }
-        )
+            )
+        }
 
         context("the right sub config is passed to the rule") {