--- a/detekt-metrics/build.gradle.kts
+++ b/detekt-metrics/build.gradle.kts
@@ -1,4 +1,5 @@
 dependencies {
     api(project(":detekt-api"))
-    testImplementation(project(":detekt-test"))
+    testImplementation(project(":detekt-test-utils"))
+    testImplementation(testFixtures(project(":detekt-api")))
 }