--- a/detekt-report-txt/build.gradle.kts
+++ b/detekt-report-txt/build.gradle.kts
@@ -1,4 +1,4 @@
 dependencies {
     implementation(project(":detekt-api"))
-    testImplementation(project(":detekt-test"))
+    testImplementation(testFixtures(project(":detekt-api")))
 }