--- a/detekt-test-utils/src/main/kotlin/io/gitlab/arturbosch/detekt/rules/KotlinEnvironmentTestSetup.kt
+++ b/detekt-test-utils/src/main/kotlin/io/gitlab/arturbosch/detekt/rules/KotlinEnvironmentTestSetup.kt
@@ -6,10 +6,8 @@ import org.spekframework.spek2.dsl.Root
 import org.spekframework.spek2.lifecycle.CachingMode
 
 fun Root.setupKotlinEnvironment() {
-    val wrapper = createEnvironment()
+    val wrapper by memoized(CachingMode.SCOPE, { createEnvironment() }, { it.dispose() })
 
     @Suppress("UNUSED_VARIABLE") // name is used for delegation
     val env: KotlinCoreEnvironment by memoized(CachingMode.EACH_GROUP) { wrapper.env }
-
-    afterGroup { wrapper.dispose() }
 }