--- a/gradle.properties
+++ b/gradle.properties
@@ -2,6 +2,7 @@ kotlin.code.style=official
 systemProp.sonar.host.url=http://localhost:9000
 systemProp.file.encoding=UTF-8
 org.gradle.parallel=true
+org.gradle.workers.max=4
 org.gradle.caching=true
 # Needed for https://github.com/gradle/gradle/issues/11412
 systemProp.org.gradle.internal.publish.checksums.insecure=true