… detekt-test/src/main/kotlin/io/gitlab/arturbosch/detekt/test/TestDetektion.kt
… detekt-test/src/main/kotlin/io/gitlab/arturbosch/detekt/test/TestFactory.kt