Findings:
- recent spek discovery improvements already help a bit (3s spek vs 1s jupiter -> 2s spek vs 1s jupiter for report-xml module)
- there are still some concurrent running things we don't understand
  - the more parallelization we use the more the test discovery rises
    - using a maximum of 4 workers help a lot
- using Gradle test fixtures improving the test discovery the most! **Edit: not true, they start instantly just in IntelliJ**
  - ~module report-txt without the `detekt-test` dependency starts *instantly*!!!~
  - ~does this mean spek tries to discover to much? still running through all jars/the whole classpath??~
  - ~have we setup the KtTestCompiler and ScriptEngine wrongly??~
     - ~both classes are only used in methods and Kotlin `object`s guarantees us singletons~
     - ~I've tried to remove the `object` declarations and using custom `lazy` inits -> no improvements~

References:
- Closes https://github.com/detekt/detekt/issues/2902
- https://github.com/spekframework/spek/pull/907
- https://github.com/spekframework/spek/pull/905