@arturbosch 2020-08-01
Even with test fixtures the test discovery is noticeable when using `--parallel` (threads are starving)

![2020-08-01-210022_444x411_scrot](https://user-images.githubusercontent.com/20924106/89108655-640d3400-d43a-11ea-909f-b0c0cec6f3e2.png)