@schalkms 2020-08-01
There are 2 hard problems in computer science: threads, naming things, and off-by-1 errors.