@arturbosch 2020-08-01
Did some profiling and spek is fast enough. Classgraph single-threaded works too.
I suspect 5 coroutine threads x #cores to be the reason (spek seems to be concurrent by default).
Junit Jupiter runs the tests sequentially (https://junit.org/junit5/docs/current/user-guide/#writing-tests-parallel-execution).
Which makes sense when using build tool parallelization.