@raniejade 2020-08-02
One thing weird I found is when running `detekt-core:test` as itself, discovery is fast around 5 seconds, but when you run all the tests together - discovery goes up to 15s!