@raniejade 2020-08-02
Changing this made the discrepancy smaller from 5s (alone) -> 7s (all tests)

![Screenshot from 2020-08-02 12-16-37](https://user-images.githubusercontent.com/1041919/89113910-120bf300-d4ba-11ea-8ff0-27d35a65a31f.png)

It was declared like this before:

```kotlin
        val finding = mockk<Finding>()
        every { finding.id }.returns("LongParameterList")
        every { finding.signature }.returns("Signature")

        val result = TestDetektion(finding)
```

So it was building the mock during discovery phase.