@raniejade 2020-08-02
One improvement I could suggest is avoid doing heavy stuff in the group scopes, use a combination of fixtures (beforeEachTest, beforeEachGroup, etc ..) and memoized to delay calculation values used during the execution phase.