@raniejade 2020-08-02
I've also made a change to `setupKotlinEnvironment` to delay calculation of the environment during the execution phase.

```kotlin
fun Root.setupKotlinEnvironment() {
    val wrapper by memoized(CachingMode.SCOPE, { createEnvironment() }, { it.dispose() } )

    @Suppress("UNUSED_VARIABLE") // name is used for delegation
    val env: KotlinCoreEnvironment by memoized(CachingMode.EACH_GROUP) { wrapper.env }
}

```