@raniejade 2020-08-02
After a couple of testing from this branch, the biggest difference I can see is when making sure nothing heavy is calculated in the discovery phase. I've tried making cg scanning and execution sequential but the difference is very negligible 1-2 seconds.

[1] I made changes to your test (2m 9.625s): https://scans.gradle.com/s/hvit4fbjj54j6/tests?collapse-all

vs

[2] No changes (2m 37.258s): https://scans.gradle.com/s/k3cwnp74ersju/tests?collapse-all

From what I can see the individual test timings in Gradle's scan doesn't account for discovery time - but just execution time.