@arturbosch 2020-08-02
@raniejade thank you very much for taking the time to help us out!
I've incorporated your suggestions.

I also removed the coroutine usage but the time between starting the test task and executing the tests is nearly the same (1-2s as you mentioned). 

So the only thing we can do is to spop introducing new modules which need the kotlin compiler or mockk :/.