@raniejade 2020-08-02
> So the only thing we can do is to stop introducing new modules which need the kotlin compiler or mockk :/.

I think if we can just make sure that no heavy initialization happens in the discovery phase, will help improve discovery. There are probably other areas in your tests that are doing some initialization in the discovery phase, fixing them might help improve test speeds.